Renamed the "types" file to "_types" since it was conflicting with the python type files on android.
